+++ FallRiver/FallRiver/LoseMenuState.cpp
@@ -152,7 +152,7 @@ void LoseMenuState::Render()
 	else if(curPos == 500)
 	{ExitRect.left = 1535; ExitRect.right = ExitRect.left+287;}
 
-	pVM->DrawStaticTexture(tempLoseID, 170, curPos-50, 1.0f, 1.0f, &BloodRect);
+	pVM->DrawStaticTexture(tempLoseID, 170, (float)curPos-50, 1.0f, 1.0f, &BloodRect);
 	pVM->DrawStaticTexture(tempLoseID, 350, 400,  0.5f, 0.9f, &resumeRect);
 	pVM->DrawStaticTexture(tempLoseID, 370, 500,  0.5f, 0.9f, &ExitRect);
 }
